[async] Split gamma interactions from transport kernel.
Create three new queues that contain the particles that are undergoing
interactions.

Not that this commit changes some simulation results, since the gamma transport
and interactions are now happening in different kernels. This breaks
some computations such as energy or position update in two, preventing
the compiler from using fma instructions, leading to slightly different
positions or energies in some events.